Rename `command_dict_creator` to `command_creator`
The fact it creates a dict is not relevant to its callers
